+++ src/main/java/es/upm/grise/profundizacion/HandleDocuments/ReflectionWrapper.java
@@ -1,0 +1,7 @@
+package es.upm.grise.profundizacion.HandleDocuments;
+
+public class ReflectionWrapper {
+    public Class<?> findClass(String className) throws ClassNotFoundException {
+        return Class.forName(className);
+    }
+}